+++ src/styles/components/CameraPage.css
@@ -1,0 +1,258 @@
+/* Estilos para la página de captura */
+.capture-page {
+  background: var(--white);
+  border-radius: var(--border-radius);
+  padding: 40px;
+  box-shadow: var(--shadow-medium);
+  max-width: none;
+}
+
+.webcam-wrapper {
+  position: relative;
+  width: 100%;
+  max-width: 640px;
+  margin: 0 auto;
+}
+
+.webcam {
+  width: 100%;
+  max-width: 640px;
+  border-radius: 8px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+.captured-image {
+  width: 100%;
+  max-width: 640px;
+  border-radius: 8px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  object-fit: contain;
+}
+
+/* Estilos para la guía de colocación de ID */
+.webcam-container {
+  position: relative;
+  width: 100%;
+  max-width: 640px;
+  margin: 0 auto;
+}
+
+.id-guide-container {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 20;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  pointer-events: none;
+  opacity: 1;
+  transition: opacity 0.3s ease;
+}
+
+.id-guide-container.hidden {
+  opacity: 0;
+}
+
+.id-guide-overlay {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background-color: rgba(0, 0, 0, 0.7);
+  padding: 15px;
+  border-radius: 8px;
+  max-width: 90%;
+}
+
+.id-guide-image {
+  width: 100%;
+  max-width: 300px;
+  height: auto;
+  margin-bottom: 10px;
+  border: 2px solid white;
+  border-radius: 4px;
+}
+
+.id-guide-text {
+  color: white;
+  text-align: center;
+  font-size: 14px;
+  line-height: 1.3;
+}
+
+.id-guide-text p {
+  margin: 5px 0;
+}
+
+.btn-guide {
+  background-color: #ff9800;
+  color: white;
+  font-weight: bold;
+  border-radius: 50%;
+  width: 40px;
+  height: 40px;
+  padding: 0;
+  margin-left: 10px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.btn-guide:hover {
+  background-color: #f57c00;
+  transform: scale(1.05);
+}
+
+.resolution-info {
+  position: absolute;
+  bottom: 10px;
+  right: 10px;
+  background-color: rgba(0, 0, 0, 0.7);
+  color: white;
+  padding: 4px 8px;
+  border-radius: 4px;
+  font-size: 12px;
+  font-family: monospace;
+  z-index: 10;
+}
+
+.resolution-check {
+  display: inline-block;
+  margin-left: 10px;
+  background-color: var(--primary-green);
+  padding: 2px 6px;
+  border-radius: 4px;
+  font-size: 12px;
+  color: white;
+}
+
+.resolution-warning {
+  display: inline-block;
+  margin-left: 10px;
+  background-color: rgba(255, 193, 7, 0.9);
+  padding: 2px 6px;
+  border-radius: 4px;
+  font-size: 12px;
+  color: #000;
+  font-weight: bold;
+}
+
+.check-icon {
+  font-weight: bold;
+  color: #00ff00;
+}
+
+.devices-container {
+  margin: 20px 0;
+  text-align: center;
+}
+
+.device-list {
+  margin-top: 20px;
+}
+
+.device-item {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 15px;
+  margin: 10px 0;
+  background-color: var(--light-gray);
+  border-radius: var(--border-radius);
+  border: 1px solid var(--medium-gray);
+}
+
+.device-item span {
+  color: var(--text-gray);
+}
+
+.camera-container {
+  margin: 20px 0;
+  text-align: center;
+}
+
+.camera-controls {
+  margin: 15px 0;
+  display: flex;
+  justify-content: center;
+  gap: 10px;
+}
+
+.resolution-buttons {
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  gap: 10px;
+  margin-top: 15px;
+}
+
+.btn-resolution {
+  background-color: var(--medium-gray);
+  color: var(--dark-gray);
+  border: 1px solid var(--medium-gray);
+  position: relative;
+  overflow: hidden;
+}
+
+.btn-resolution:hover {
+  background-color: #d0d0d0;
+  border-color: var(--dark-gray);
+}
+
+.btn-resolution-active {
+  background-color: var(--primary-green);
+  color: var(--white);
+  border-color: var(--primary-green);
+}
+
+.btn-resolution::after {
+  content: '';
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  width: 0;
+  height: 0;
+  border-radius: 50%;
+  background: rgba(255, 255, 255, 0.3);
+  transform: translate(-50%, -50%);
+  transition: width 0.5s, height 0.5s;
+}
+
+.btn-resolution:active::after {
+  width: 100%;
+  height: 100%;
+}
+
+/* Estilos para indicadores de progreso de upload */
+.upload-progress {
+  margin: 10px 0;
+  text-align: center;
+  padding: 10px;
+}
+
+.progress-bar {
+  width: 100%;
+  height: 20px;
+  background-color: #f0f0f0;
+  border-radius: 10px;
+  overflow: hidden;
+  margin-top: 10px;
+}
+
+.progress-bar-fill {
+  height: 100%;
+  background-color: #4CAF50;
+  transition: width 0.3s ease;
+}
+
+.upload-error {
+  color: #ff0000;
+  text-align: center;
+  margin: 10px 0;
+  padding: 10px;
+  background-color: #ffebee;
+  border-radius: 5px;
+  border: 1px solid #ffcdd2;
+}
